rag-engine/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
groq
python-dotenv
numpy
pandas